Refactor CountryController: improved readability and enhance code quality

+++ BasicApiNet.Host/Controllers/CountryController.cs
@@ -2,6 +2,7 @@ using BasicApiNet.Core.Dtos;
 using BasicApiNet.Core.Models;
 using BasicApiNet.Core.Services;
 using Microsoft.AspNetCore.Mvc;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace BasicApiNet.Host.Controllers;
 
@@ -17,16 +18,26 @@ public class CountryController : ControllerBase
     }
 
     [HttpGet]
-    public ActionResult<IEnumerable<Country>> GetAll()
+    [SwaggerOperation(Summary = "Get all countries", Description = "Retrieve all countries.")]
+    [SwaggerResponse(200, "Success", typeof(IEnumerable<Country>))]
+    [SwaggerResponse(500, "Internal Server Error", typeof(string))]
+    public Task<IActionResult> GetAll()
     {
         var response = _service.GetAll().ToList();
-        return Ok(response);
+        return Task.FromResult<IActionResult>(Ok(response));
     }
 
     [HttpPost]
-    public async Task<ActionResult> Create(Country country,
-        CancellationToken cancellationToken)
+    [SwaggerOperation(Summary = "Create country", Description = "Create a country.")]
+    [SwaggerResponse(200, "Success", typeof(bool))]
+    [SwaggerResponse(400, "Bad Request", typeof(string))]
+    [SwaggerResponse(500, "Internal Server Error", typeof(string))]
+    public async Task<IActionResult> Create(Country country)
     {
+        if (string.IsNullOrEmpty(country.Name) || string.IsNullOrWhiteSpace(country.Name))
+        {
+            return BadRequest("Invalid country");
+        }
         try
         {
             await _service.CreateAsync(country);
@@ -34,33 +45,48 @@ public class CountryController : ControllerBase
         }
         catch (Exception e)
         {
-            return Ok(false);
+            return StatusCode(500, $"Internal server error: {e.Message}");
         }
     }
 
     [HttpPost]
-    public async Task<ActionResult> CreateWithCities([FromBody] CountryWithCitiesRequest request,
+    [SwaggerOperation(Summary = "Create countries with their cities(optional)", Description = "Create a country with its cities(optional).")]
+    [SwaggerResponse(200, "Success", typeof(bool))]
+    [SwaggerResponse(400, "Bad Request", typeof(string))]
+    [SwaggerResponse(500, "Internal Server Error", typeof(string))]
+    public async Task<IActionResult> CreateWithCities([FromBody] CountryWithCitiesRequest request,
         CancellationToken cancellationToken)
     {
+        if (string.IsNullOrEmpty(request.Country.Name) || string.IsNullOrWhiteSpace(request.Country.Name))
+        {
+            return BadRequest("Invalid country");
+        }
         try
         {
             //La logicas de negocio para crear un pais con ciudades deberia ir en un manager service.
             request.Country.Cities.AddRange(request.Cities);
 
-            _service.CreateAsync(request.Country);
+            await _service.CreateAsync(request.Country);
 
             return Ok(true);
         }
         catch (Exception e)
         {
-            return Ok(false);
+            return StatusCode(500, $"Internal server error: {e.Message}");
         }
     }
 
     [HttpPut]
-    public async Task<ActionResult> Update(Country country,
-        CancellationToken cancellationToken)
+    [SwaggerOperation(Summary = "Update country", Description = "Update a country.")]
+    [SwaggerResponse(200, "Success", typeof(Country))]
+    [SwaggerResponse(400, "Bad Request", typeof(string))]
+    [SwaggerResponse(500, "Internal Server Error", typeof(string))]
+    public async Task<IActionResult> Update(Country country)
     {
+        if (country.Id == 0 || country.Id < 0 || string.IsNullOrEmpty(country.Name) || string.IsNullOrWhiteSpace(country.Name))
+        {
+               return BadRequest("Invalid country");
+        }
         try
         {
             var countryUpdated = await _service.UpdateAsync(country);
@@ -68,22 +94,43 @@ public class CountryController : ControllerBase
         }
         catch (Exception e)
         {
-            return Ok(false);
+            return StatusCode(500, $"Internal server error: {e.Message}");
         }
     }
 
     [HttpGet("{id}")]
-    public async Task<ActionResult<Country>> GetById(int id,
-        CancellationToken cancellationToken)
+    [SwaggerOperation(Summary = "Get country by id", Description = "Retrieve a country by id.")]
+    [SwaggerResponse(200, "Success", typeof(Country))]
+    [SwaggerResponse(400, "Bad Request", typeof(string))]
+    [SwaggerResponse(500, "Internal Server Error", typeof(string))]
+    public async Task<IActionResult> GetCountry(int id)
     {
-        var response = _service.FinByIdAsync(id);
-        return Ok(response);
+        if (id is 0 or < 0)
+        {
+            return BadRequest("Invalid id");
+        }
+        try
+        {
+            var response = await _service.FinByIdAsync(id);
+            return Ok(response);
+        }
+        catch (Exception e)
+        {
+            return StatusCode(500, $"Internal server error: {e.Message}");
+        }
+
     }
 
     [HttpDelete("{id}")]
-    public ActionResult Delete(int id,
-        CancellationToken cancellationToken)
+    [SwaggerOperation(Summary = "Delete country by id", Description = "Delete a country by id.")]
+    [SwaggerResponse(200, "Success", typeof(bool))]
+    [SwaggerResponse(500, "Internal Server Error", typeof(string))]
+    public IActionResult Delete(int id)
     {
+        if (id is 0 or < 0)
+        {
+            return BadRequest("Invalid id");
+        }
         try
         {
             _service.DeleteById(id);
@@ -91,7 +138,7 @@ public class CountryController : ControllerBase
         }
         catch (Exception e)
         {
-            return Ok(false);
+            return StatusCode(500, $"Internal server error: {e.Message}");
         }
     }
 }
